+++ src/Bundles/Scanner/TWIG/TwigScanner.php
@@ -1,0 +1,36 @@
+<?php
+
+namespace PHPUnuhi\Bundles\Twig;
+
+class TwigScanner implements ScannerInterface
+{
+
+    /**
+     * @return string
+     */
+    public function getStorageName(): string
+    {
+        return 'twig';
+    }
+
+    /**
+     * {{ 'header.example' | trans }}
+     *
+     * @param string $key
+     * @param string $content
+     * @return bool
+     */
+    public function findKey(string $key, string $content): bool
+    {
+        $content = str_replace(" ", '', $content);
+        $content = str_replace('"', "'", $content);
+
+        $pattern = '/{{\s*\'?' . preg_quote($key, '/') . '\'?\s*\|\s*.*trans.*\s*}}/';
+
+        $matches = [];
+
+        preg_match($pattern, $content, $matches);
+
+        return (!empty($matches));
+    }
+}